+++ duckscript_sdk/src/sdk/std/env/cpu_count/mod.rs
@@ -1,0 +1,42 @@
+use crate::utils::pckg;
+use duckscript::types::command::{Command, CommandResult};
+use num_cpus;
+
+#[cfg(test)]
+#[path = "./mod_test.rs"]
+mod mod_test;
+
+#[derive(Clone)]
+pub(crate) struct CommandImpl {
+    package: String,
+}
+
+impl Command for CommandImpl {
+    fn name(&self) -> String {
+        pckg::concat(&self.package, "GetCpuCount")
+    }
+
+    fn aliases(&self) -> Vec<String> {
+        vec!["cpu_count".to_string(), "get_cpu_count".to_string()]
+    }
+
+    fn help(&self) -> String {
+        include_str!("help.md").to_string()
+    }
+
+    fn clone_and_box(&self) -> Box<dyn Command> {
+        Box::new((*self).clone())
+    }
+
+    fn run(&self, _arguments: Vec<String>) -> CommandResult {
+        let num = num_cpus::get();
+
+        CommandResult::Continue(Some(num.to_string()))
+    }
+}
+
+pub(crate) fn create(package: &str) -> Box<dyn Command> {
+    Box::new(CommandImpl {
+        package: package.to_string(),
+    })
+}